Mock timestamper working in FIFO mode
